requirements.txt
nltk==3.7
pandas==1.4.2
pytorch_lightning==1.6.3
PyYAML==6.0
scikit_learn==1.1.1
sentence_transformers==2.2.0
torch==1.11.0
transformers==4.19.2
